pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "immunotype"
version = "0.0.1"
description = "Peptide-based HLA typing from immunopeptidomics data"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "immunotype contributors"}
]
keywords = ["immunopeptidomics", "HLA", "typing", "peptides", "machine learning", "bioinformatics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "torch>=2.0.0",
    "torch-geometric>=2.3.0",
    "tqdm>=4.64.0",
    "rich-click>=1.6.0",
    "gradio>=4.0.0",
    "seaborn>=0.11.0",
    "pathlib",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
    "types-tqdm",
    "types-seaborn",
]
server = [
    "uvicorn[standard]>=0.24.0",
    "fastapi>=0.104.0",
]
all = [
    "immunotype[dev,server]"
]

[project.urls]
Homepage = "https://github.com/immunotype/immunotype"
Documentation = "https://github.com/immunotype/immunotype#readme"
Repository = "https://github.com/immunotype/immunotype.git"
Issues = "https://github.com/immunotype/immunotype/issues"

[project.scripts]
immunotype = "immunotype.cli:main"
immunotype-web = "immunotype.web:app.launch"
immunotype-server = "immunotype.server:app"

# Tool configurations
[tool.hatch.build.targets.wheel]
packages = ["src/immunotype"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=immunotype",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torch_geometric.*",
    "seaborn.*",
    "gradio.*",
    "tqdm.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "B904", # raise from
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["E501", "D"]

[tool.ruff.isort]
known-first-party = ["immunotype"]
force-single-line = true